Improve and fix typing of LLM ETH send flow
