+++ apps/ledger-live-mobile/src/families/ethereum/EthereumCustomFees/Ethereum1559CustomFees.tsx
@@ -2,13 +2,11 @@ import React, { useState, memo, useMemo, useCallback } from "react";
 import { formatCurrencyUnit } from "@ledgerhq/live-common/currencies/formatCurrencyUnit";
 import useBridgeTransaction from "@ledgerhq/live-common/bridge/useBridgeTransaction";
 import { getGasLimit } from "@ledgerhq/live-common/families/ethereum/transaction";
+import { useTheme } from "@react-navigation/native";
 import { Transaction } from "@ledgerhq/live-common/families/ethereum/types";
 import { getMainAccount } from "@ledgerhq/live-common/account/helpers";
-import { inferDynamicRange, Range } from "@ledgerhq/live-common/range";
 import { getAccountBridge } from "@ledgerhq/live-common/bridge/index";
-import { useRoute, useTheme } from "@react-navigation/native";
 import { Account, AccountLike } from "@ledgerhq/types-live";
-import { getEnv } from "@ledgerhq/live-common/env";
 import { StyleSheet, View } from "react-native";
 import { BigNumber } from "bignumber.js";
 import { Trans } from "react-i18next";
@@ -17,45 +15,8 @@ import LText from "../../../components/LText";
 import Button from "../../../components/Button";
 import EthereumGasLimit from "../SendRowGasLimit";
 import EditFeeUnitEthereum from "../EditFeeUnitEthereum";
-import { RouteParams } from "../../../screens/SendFunds/04-Summary";
 import SectionSeparator from "../../../components/SectionSeparator";
-
-const TWENTY_GWEI = new BigNumber(10e9);
-
-const defaultMaxPriorityFeeRange = inferDynamicRange(TWENTY_GWEI); // 0 - 20 Gwei
-const defaultMaxFeePerGasRange = inferDynamicRange(TWENTY_GWEI); // 0 - 20 Gwei
-
-const inferMaxPriorityFeeRange = (
-  networkInfo: Transaction["networkInfo"],
-): Range => {
-  if (!networkInfo?.maxPriorityFeePerGas || !networkInfo?.nextBaseFeePerGas)
-    return defaultMaxPriorityFeeRange;
-
-  return inferDynamicRange(networkInfo.maxPriorityFeePerGas.initial, {
-    minValue: getEnv("EIP1559_MINIMUM_FEES_GATE")
-      ? networkInfo.maxPriorityFeePerGas.min.times(
-          getEnv("EIP1559_PRIORITY_FEE_LOWER_GATE"),
-        )
-      : new BigNumber(0),
-    maxValue: networkInfo.maxPriorityFeePerGas.max,
-  });
-};
-
-const inferMaxFeeRange = (networkInfo: Transaction["networkInfo"]): Range => {
-  if (!networkInfo?.maxPriorityFeePerGas || !networkInfo?.nextBaseFeePerGas)
-    return defaultMaxFeePerGasRange;
-
-  const { maxPriorityFeePerGas, nextBaseFeePerGas } = networkInfo;
-  const amplifiedBaseFee = nextBaseFeePerGas.times(2); // ensuring 6 blocks of base fee
-
-  return inferDynamicRange(
-    amplifiedBaseFee.plus(maxPriorityFeePerGas.initial),
-    {
-      minValue: new BigNumber(0),
-      maxValue: amplifiedBaseFee.plus(maxPriorityFeePerGas.max),
-    },
-  );
-};
+import { inferMaxFeeRange, inferMaxPriorityFeeRange } from "./utils";
 
 type Props = {
   account: AccountLike;
@@ -70,7 +31,6 @@ const Ethereum1559CustomFees = ({
   onValidateFees,
   transaction: originalTransaction,
 }: Props) => {
-  const route = useRoute();
   const { colors } = useTheme();
 
   invariant(originalTransaction.family === "ethereum", "not ethereum family");
@@ -228,7 +188,6 @@ const Ethereum1559CustomFees = ({
           account={account}
           parentAccount={parentAccount}
           transaction={transaction as Transaction}
-          route={route as { params: RouteParams }}
           gasLimit={gasLimit}
           setGasLimit={setGasLimit}
         />
