+++ apps/ledger-live-mobile/src/families/ethereum/EthereumCustomFees/index.tsx
@@ -7,16 +7,22 @@ import { getAccountBridge } from "@ledgerhq/live-common/bridge/index";
 import { getMainAccount } from "@ledgerhq/live-common/account/helpers";
 import { Transaction } from "@ledgerhq/live-common/families/ethereum/types";
 import { EIP1559ShouldBeUsed } from "@ledgerhq/live-common/families/ethereum/transaction";
+import { SendFundsNavigatorStackParamList } from "../../../components/RootNavigator/types/SendFundsNavigator";
 import { accountScreenSelector } from "../../../reducers/accounts";
 import EthereumLegacyCustomFees from "./EthereumLegacyCustomFees";
-import { Props as RouteParams } from "../EthereumFeesStrategy";
 import Ethereum1559CustomFees from "./Ethereum1559CustomFees";
+import {
+  BaseComposite,
+  StackNavigatorProps,
+} from "../../../components/RootNavigator/types/helpers";
+import { ScreenName } from "../../../const";
 
-type Props = {
-  route: {
-    params: RouteParams;
-  };
-};
+type Props = BaseComposite<
+  StackNavigatorProps<
+    SendFundsNavigatorStackParamList,
+    ScreenName.EthereumCustomFees
+  >
+>;
 
 const options = {
   title: <Trans i18nKey="send.summary.fees" />,
@@ -24,7 +30,7 @@ const options = {
 };
 
 export default function EthereumCustomFees({ route }: Props) {
-  const { setTransaction, transaction, status } = route.params;
+  const { setTransaction, transaction } = route.params;
   const { account, parentAccount } = useSelector(accountScreenSelector(route));
   const navigation = useNavigation();
   invariant(account, "no account found");
@@ -48,7 +54,6 @@ export default function EthereumCustomFees({ route }: Props) {
       parentAccount={parentAccount}
       transaction={transaction}
       onValidateFees={onValidateFees}
-      status={status}
     />
   ) : (
     <EthereumLegacyCustomFees
@@ -56,7 +61,6 @@ export default function EthereumCustomFees({ route }: Props) {
       parentAccount={parentAccount}
       transaction={transaction}
       onValidateFees={onValidateFees}
-      status={status}
     />
   );
 }
