+++ apps/ledger-live-mobile/src/families/ethereum/SendRowsFee.tsx
@@ -1,47 +1,13 @@
 import React from "react";
-import type { Account, AccountLike } from "@ledgerhq/types-live";
-import type { Transaction } from "@ledgerhq/live-common/generated/types";
-import { CompositeScreenProps } from "@react-navigation/native";
+import { Transaction } from "@ledgerhq/live-common/families/ethereum/types";
 import EthereumFeesStrategy from "./EthereumFeesStrategy";
-import type { StackNavigatorProps } from "../../components/RootNavigator/types/helpers";
-import type { SendFundsNavigatorStackParamList } from "../../components/RootNavigator/types/SendFundsNavigator";
-import type { BaseNavigatorStackParamList } from "../../components/RootNavigator/types/BaseNavigator";
-import { ScreenName } from "../../const";
-import type { LendingEnableFlowParamsList } from "../../components/RootNavigator/types/LendingEnableFlowNavigator";
-import type { LendingSupplyFlowNavigatorParamList } from "../../components/RootNavigator/types/LendingSupplyFlowNavigator";
-import type { LendingWithdrawFlowNavigatorParamList } from "../../components/RootNavigator/types/LendingWithdrawFlowNavigator";
-import type { SignTransactionNavigatorParamList } from "../../components/RootNavigator/types/SignTransactionNavigator";
-import type { SwapNavigatorParamList } from "../../components/RootNavigator/types/SwapNavigator";
+import { SendRowsFeeProps } from "./types";
 
-type Props = {
-  transaction: Transaction;
-  account: AccountLike;
-  parentAccount?: Account | null;
-  setTransaction: (..._: Array<Transaction>) => void;
-} & CompositeScreenProps<
-  | StackNavigatorProps<
-      SendFundsNavigatorStackParamList,
-      ScreenName.SendSummary
-    >
-  | StackNavigatorProps<
-      SignTransactionNavigatorParamList,
-      ScreenName.SignTransactionSummary
-    >
-  | StackNavigatorProps<
-      LendingEnableFlowParamsList,
-      ScreenName.LendingEnableSummary
-    >
-  | StackNavigatorProps<
-      LendingSupplyFlowNavigatorParamList,
-      ScreenName.LendingSupplySummary
-    >
-  | StackNavigatorProps<
-      LendingWithdrawFlowNavigatorParamList,
-      ScreenName.LendingWithdrawSummary
-    >
-  | StackNavigatorProps<SwapNavigatorParamList, ScreenName.SwapSelectFees>,
-  StackNavigatorProps<BaseNavigatorStackParamList>
->;
-export default function EthereumSendRowsFee(props: Props) {
-  return <EthereumFeesStrategy {...props} />;
+export default function EthereumSendRowsFee({
+  transaction,
+  ...props
+}: SendRowsFeeProps) {
+  return (
+    <EthereumFeesStrategy transaction={transaction as Transaction} {...props} />
+  );
 }
